requirements.txt
fastapi
uvicorn
paho-mqtt
python-kasa
requests
pyyaml
